clean up extension array module
